requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
gunicorn==21.2.0
Markdown==3.6
packaging==23.2
pillow==10.2.0
sqlparse==0.4.4
tzdata==2023.4
